backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.2
python-multipart==0.0.6
pillow==9.5.0
numpy==1.24.3
scikit-image==0.20.0
scipy==1.10.1
torch==2.0.1
torchvision==0.15.2
pytest==7.3.1
httpx==0.24.1
python-dotenv==1.0.0
opencv-python==4.7.0.72
matplotlib==3.7.1